[#93] Add the ability to skip building the application image
If `build.s2i.buildApplication` is set to `false`, the resources to
create the applicaition image are not generated.
In that case, the builder-artifacts image will take over the name of the
Helm release (as it is the only produced image).

Deployment of the builder image can still happen and needs to be
explicitely disabled with `deploy.enabled` set to `false` if needs be.

This fixes #93

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>